+++ src/dune_rules/lib_rules.ml
@@ -20,7 +20,8 @@ let build_lib (lib : Library.t) ~native_archives ~sctx ~expander ~flags ~dir
   Memo.Result.iter (Context.compiler ctx mode) ~f:(fun compiler ->
       let target = Library.archive lib ~dir ~ext:(Mode.compiled_lib_ext mode) in
       let stubs_flags =
-        let lib_archive, foreign_archives = Library.foreign_archives lib in
+        let lib_archive = Library.stubs_archive lib in
+        let foreign_archives = Library.foreign_archives lib in
         let make_args ~stub_mode archive =
           let lname =
             "-l"
@@ -290,7 +291,8 @@ let build_stubs lib ~cctx ~dir ~expander ~requires ~dir_contents
     if
       Mode.Dict.Set.to_list modes
       |> List.for_all ~f:(fun mode ->
-             List.is_empty @@ Mode.Map.Multi.for_only o_files mode)
+             List.is_empty
+             @@ Mode.Map.Multi.for_only ~and_all:false o_files mode)
     then
       (* if stubs are not mode dependent *)
       let o_files = for_all_modes in
@@ -300,7 +302,9 @@ let build_stubs lib ~cctx ~dir ~expander ~requires ~dir_contents
       let modes =
         Mode.Dict.Set.to_list modes
         |> List.map ~f:(fun mode ->
-               let o_files_for_mode = Mode.Map.Multi.for_only o_files mode in
+               let o_files_for_mode =
+                 Mode.Map.Multi.for_only ~and_all:false o_files mode
+               in
                ( List.rev_append for_all_modes o_files_for_mode
                , Mode.Select.Only mode ))
       in